Implementing algorithm
Used a insertion sort function to put them in an acceptable install
order. Created function to return the package name for output. Created
function that checks for loop. Basically since we know the sort works,
we know that any item that's not "sorted" (it's listed after its
dependent) must necessarily be part of a loop.